% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjoint_rImp.R
\name{conjoint_rImp}
\alias{conjoint_rImp}
\title{Conjoint lm}
\usage{
conjoint_rImp(formula, data)
}
\arguments{
\item{formula}{an object of class formula (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which lm is called}
}
\value{
dataframe
}
\description{
Return data table that remove respondents provided different answer in the same questions combination
}
\examples{
\dontrun{

remove_contradict(experiment_data, rtb, answer, response_id)

}

}
