% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_distribution.R
\name{score_distribution}
\alias{score_distribution}
\title{Calculate the distribution of score by variables}
\usage{
score_distribution(data, ..., score)
}
\arguments{
\item{data}{a data frame}

\item{...}{variables or computations to group by.}

\item{score}{variable which is used for distribution}
}
\value{
dataframe
}
\description{
Return data table showing distribution of score by variables
}
\examples{
\dontrun{

score_distribution(experiment_data, rtb, answer)

}

}
