% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_distribution.R
\name{variable_distribution}
\alias{variable_distribution}
\title{Calculate number of count of selected dataset of the variable}
\usage{
variable_distribution(data, ...)
}
\arguments{
\item{data}{a data frame}

\item{...}{variables or computations to group by.}
}
\value{
dataframe
}
\description{
Return data table showing number of count by the variable. You can only input 1 variable of the original dataset
}
\examples{
\dontrun{

Variable_distribution(data = experiment_data, price, offer)

}

}
